+++ packages/terra-form-select/CHANGELOG.md
@@ -4,10 +4,18 @@ ChangeLog
 Unreleased
 ----------
 
+6.7.0 - (June 9, 2020)
+------------------
+### Changed
+* Minor file changes to comply with eslint rules
+* Update wdio test configuration to use root wdio config
+
 6.6.0 - (June 2, 2020)
 ------------------
 ### Changed
 * Apply themes by context and directly include packaged themes.
+* Updated wdio test to support testing against firefox
+* Disabled wdio axe color-contrast testing on placeholder examples for IE testing. This will be re-enabled when https://github.com/cerner/terra-core/issues/2992 is resolved.
 
 6.5.0 - (May 19, 2020)
 ------------------